refactor(pages/home): extract `<GradientFeatures />`
