+++ components/home/GradientFeatures.test.tsx
@@ -1,0 +1,14 @@
+import { render, screen } from "@testing-library/react";
+import { GradientFeatures } from "components/home/GradientFeatures";
+
+describe("<GradientFeatures />", () => {
+  it("renders", () => {
+    // Arrange
+
+    // Act
+    const { asFragment } = render(<GradientFeatures />);
+
+    // Assert
+    expect(asFragment()).toMatchSnapshot();
+  });
+});
