+++ components/home/GradientFeatures.tsx
@@ -1,0 +1,101 @@
+import {
+  ChatAltIcon,
+  DocumentReportIcon,
+  HeartIcon,
+  InboxIcon,
+  PencilAltIcon,
+  ReplyIcon,
+  TrashIcon,
+  UsersIcon,
+} from "@heroicons/react/outline";
+
+const features = [
+  {
+    name: "Unlimited Inboxes",
+    description:
+      "Ac tincidunt sapien vehicula erat auctor pellentesque rhoncus. Et magna sit morbi lobortis.",
+    icon: InboxIcon,
+  },
+  {
+    name: "Manage Team Members",
+    description:
+      "Ac tincidunt sapien vehicula erat auctor pellentesque rhoncus. Et magna sit morbi lobortis.",
+    icon: UsersIcon,
+  },
+  {
+    name: "Spam Report",
+    description:
+      "Ac tincidunt sapien vehicula erat auctor pellentesque rhoncus. Et magna sit morbi lobortis.",
+    icon: TrashIcon,
+  },
+  {
+    name: "Compose in Markdown",
+    description:
+      "Ac tincidunt sapien vehicula erat auctor pellentesque rhoncus. Et magna sit morbi lobortis.",
+    icon: PencilAltIcon,
+  },
+  {
+    name: "Team Reporting",
+    description:
+      "Ac tincidunt sapien vehicula erat auctor pellentesque rhoncus. Et magna sit morbi lobortis.",
+    icon: DocumentReportIcon,
+  },
+  {
+    name: "Saved Replies",
+    description:
+      "Ac tincidunt sapien vehicula erat auctor pellentesque rhoncus. Et magna sit morbi lobortis.",
+    icon: ReplyIcon,
+  },
+  {
+    name: "Email Commenting",
+    description:
+      "Ac tincidunt sapien vehicula erat auctor pellentesque rhoncus. Et magna sit morbi lobortis.",
+    icon: ChatAltIcon,
+  },
+  {
+    name: "Connect with Customers",
+    description:
+      "Ac tincidunt sapien vehicula erat auctor pellentesque rhoncus. Et magna sit morbi lobortis.",
+    icon: HeartIcon,
+  },
+];
+
+export function GradientFeatures() {
+  return (
+    <div className="bg-gradient-to-r from-purple-800 to-indigo-700">
+      <div className="max-w-4xl px-4 py-16 mx-auto sm:px-6 sm:pt-20 sm:pb-24 lg:max-w-7xl lg:pt-24 lg:px-8">
+        <h2 className="text-3xl font-extrabold tracking-tight text-white">
+          Inbox support built for efficiency
+        </h2>
+        <p className="max-w-3xl mt-4 text-lg text-purple-200">
+          Ac tincidunt sapien vehicula erat auctor pellentesque rhoncus. Et
+          magna sit morbi lobortis. Blandit aliquam sit nisl euismod mattis in.
+        </p>
+        <div className="mt-12 grid grid-cols-1 gap-x-6 gap-y-12 sm:grid-cols-2 lg:mt-16 lg:grid-cols-4 lg:gap-x-8 lg:gap-y-16">
+          {features.map((feature) => (
+            <div key={feature.name}>
+              <div>
+                <span className="flex items-center justify-center w-12 h-12 bg-white rounded-md bg-opacity-10">
+                  <feature.icon
+                    className="w-6 h-6 text-white"
+                    aria-hidden="true"
+                  />
+                </span>
+              </div>
+              <div className="mt-6">
+                <h3 className="text-lg font-medium text-white">
+                  {feature.name}
+                </h3>
+                <p className="mt-2 text-base text-purple-200">
+                  {feature.description}
+                </p>
+              </div>
+            </div>
+          ))}
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export default GradientFeatures;
